services/model_training/requirements.txt
feast[postgres]
pandas
scikit-learn
psycopg2-binary
joblib